+++ Layer.h
@@ -7,7 +7,7 @@
 class Layer : public std::vector < Neuron >
 {
 public:
-	Layer(unsigned int numNeurons);
+	Layer(size_t numNeurons);
 
 	void setOutputValues(const std::vector<double> & outputValues);
 	void feedForward(const Layer &inputLayer);
